Code cleanup, notifications support variable times, allow scheduling of multiple notifications

+++ ShiftSix/app/src/main/java/com/example/shiftsix/INotificationScheduleListener.java
@@ -1,5 +1,5 @@
 package com.example.shiftsix;
 
 public interface INotificationScheduleListener {
-    public void scheduleNotification(String content, int delay);
+    public void scheduleNotification(String content, long delay);
 }
